feat: implement report creation and details functionality, enhance UI components, and refactor state management

+++ lib/features/vehicle_details_after_add/data/model/review_model.dart
@@ -21,7 +21,7 @@ class ReviewModel {
       id: json[ApiKey.getVehicleReviewId],
       reviewer: UserModel.fromJson(json[ApiKey.getVehicleReviewer]),
       createdAt: json[ApiKey.getVehicleCreatedAt],
-      rate: json[ApiKey.getVehicleRate],
+      rate: json[ApiKey.getVehicleRate].toDouble(),
       comments: json[ApiKey.getVehicleComments],
     );
   }
